Project/app/requirements.txt
flask
keras
numpy
opencv-python
scikit-image
matplotlib
Pillow
tensorflow-gpu